backend/requirements.txt
nltk
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pytest
requests
websockets
python-socketio[asyncio_client]
python-socketio[asgi]
openai
boto3
google-auth
pytest
pytest-asyncio
httpx
